Fix bug: last field of each line does not end with a comma
The regular expression required a comma at the end of each field;
as a result, the last field in each record never matched.

I updated the code to use the same technique as for end of lines,
using the regular expression to match the separator rather than values,
and the non-matching substring to extract the values.